ENH: Adds self update to the scene objects
1. Adds self update to the scene objects. The objects now can update themselves instead of explicit way of updating in the sceneManager class.
2. Clear warnings in PbdVolumeeConstraint, DeformableModel, PbdModel, NewtonMethod classes
3. Clean the main() in the sandbox example

+++ Base/Constraint/PbdConstraints/imstkPbdVolumeConstraint.h
@@ -48,8 +48,9 @@ public:
     ///
     /// \brief Initializes the volume constraint
     ///
-    void initConstraint(PositionBasedDynamicsModel& model, const unsigned int& pIdx1, const unsigned int& pIdx2,
-        const unsigned int& pIdx3, const unsigned int& pIdx4, const double k = 2.0);
+    void initConstraint(PositionBasedDynamicsModel &model, const size_t &pIdx1,
+                        const size_t &pIdx2, const size_t &pIdx3,
+                        const size_t &pIdx4, const double k = 2.0);
 
     ///
     /// \brief Solves the volume constraint
